Support for Digest Authentication
Added support for Digest Authentication (username / password). This was
necessary for integrating the SKOSjs app with a Virtuoso instance. The
user can now choose between a Token Authentication (default) and a
Digest Authentication, when logging in.

If SKOSjs is to be used against a Virtuoso instance, you need to use
Digest Authentication, against a /sparql-auth endpoint.